+++ benchmark/gzip_benchmark.dart
@@ -109,8 +109,8 @@ Future<int> _runGZipBenchmark(int dataLength) async {
     print('GZip encode/decode ${bytes.length} bytes of random data.');
     var data = GZipData(bytes);
     GZipEncodeBenchmark(data, emitter: emitter).report();
-    print(
-        'compression ratio: ${compressionRatio(bytes.length, data.bytes.length)}');
+    print('compression ratio:'
+        '${compressionRatio(bytes.length, data.bytes.length)}');
     GZipDecodeBenchmark(data, emitter: emitter).report();
     var bytesMatch = const ListEquality<int>().equals(bytes, data.bytes);
     if (bytesMatch != true) return -1;
@@ -122,8 +122,8 @@ Future<int> _runGZipBenchmark(int dataLength) async {
     print('GZip encode/decode ${bytes.length} bytes of constant data.');
     data = GZipData(bytes);
     GZipEncodeBenchmark(data, emitter: emitter).report();
-    print(
-        'compression ratio: ${compressionRatio(bytes.length, data.bytes.length)}');
+    print('compression ratio: '
+        '${compressionRatio(bytes.length, data.bytes.length)}');
     GZipDecodeBenchmark(data, emitter: emitter).report();
     bytesMatch = const ListEquality<int>().equals(bytes, data.bytes);
     return (bytesMatch != true) ? -1 : 0;