+++ lib/src/zstd/decoder.dart
@@ -6,12 +6,11 @@ import 'dart:convert';
 import 'dart:ffi';
 import 'dart:math';
 
-import 'package:es_compression/src/framework/native/filters.dart';
-
 import '../framework/buffers.dart';
 import '../framework/converters.dart';
 import '../framework/filters.dart';
 import '../framework/native/buffers.dart';
+import '../framework/native/filters.dart';
 import '../framework/sinks.dart';
 import 'ffi/constants.dart';
 import 'ffi/dispatcher.dart';